airflow-dag/requirements.txt
pandas
vaderSentiment
https://github.com/woluxwolu/twint.git
google-cloud-bigquery
google-cloud-bigquery-storage
pyarrow
twint
nest_asyncio